Turn off the buttons when not in a call or the green room.
Fixes #3

When the app detects that it's not on a page with an active call, or that
the video window is being closed, no longer draw anything on the page.

This might help with the risk of burnin and is just cleaner.

Use the "More options" menu as a proxy to detect if there's a video window,
either due to the green room or an active call.

If you do not explicitly hangup the call and navigate to a site on a
different domain the unload/beforeunload event handlers may not have enough
time to run.