Update querySelectors for new Meet style.
Now the aria-label is on the button and does not match the tooltip.  This
fix will work for english, but probably not other languages as the labels
won't match.

It's not possible in straight CSS selectors to match "button followed by div
with some text in it".  Jquery would allow it
https://api.jquery.com/contains-selector/.